Moved a lot of things to make the code cleaner
