+++ README.md
@@ -1,50 +1,41 @@
-# Welcome to your Expo app 👋
+# offcue
 
-This is an [Expo](https://expo.dev) project created with [`create-expo-app`](https://www.npmjs.com/package/create-expo-app).
+offcue is a cross‑platform reminder and scheduling app built with [Expo](https://expo.dev) and React Native. The project uses the Expo Router for navigation and provides a customizable notification system for recurring reminders.
 
-## Get started
+## Getting Started
 
-1. Install dependencies
-
+1. **Install dependencies**
    ```bash
    npm install
    ```
 
-2. Start the app
+2. **Start the development client**
+   - **iOS**
+     ```bash
+     npm run ios
+     ```
+   - **Android**
+     ```bash
+     npm run android
+     ```
 
+   Both commands launch the Expo development build in the respective simulator or emulator.
+
+3. **Optional: start Expo without building**
    ```bash
-    npx expo start
+   npm start
    ```
 
-In the output, you'll find options to open the app in a
+## Running Tests
 
-- [development build](https://docs.expo.dev/develop/development-builds/introduction/)
-- [Android emulator](https://docs.expo.dev/workflow/android-studio-emulator/)
-- [iOS simulator](https://docs.expo.dev/workflow/ios-simulator/)
-- [Expo Go](https://expo.dev/go), a limited sandbox for trying out app development with Expo
-
-You can start developing by editing the files inside the **app** directory. This project uses [file-based routing](https://docs.expo.dev/router/introduction).
-
-## Get a fresh project
-
-When you're ready, run:
+The project uses [Jest](https://jestjs.io/) for unit testing. To execute the test suite run:
 
 ```bash
-npm run reset-project
+npm test
 ```
 
-This command will move the starter code to the **app-example** directory and create a blank **app** directory where you can start developing.
+## Learn More
 
-## Learn more
+- [Expo documentation](https://docs.expo.dev/)
+- [Expo Router](https://expo.github.io/router/docs)
 
-To learn more about developing your project with Expo, look at the following resources:
-
-- [Expo documentation](https://docs.expo.dev/): Learn fundamentals, or go into advanced topics with our [guides](https://docs.expo.dev/guides).
-- [Learn Expo tutorial](https://docs.expo.dev/tutorial/introduction/): Follow a step-by-step tutorial where you'll create a project that runs on Android, iOS, and the web.
-
-## Join the community
-
-Join our community of developers creating universal apps.
-
-- [Expo on GitHub](https://github.com/expo/expo): View our open source platform and contribute.
-- [Discord community](https://chat.expo.dev): Chat with Expo users and ask questions.
