+++ components/welcome/Confirm6.tsx
@@ -54,7 +54,7 @@ export default function Confirm6({ onNext, onStartOver, reminderId }: Props) {
 
           <MenuOption
             text="☝️ Create another reminder"
-            onPress={() => router.replace("/new-reminder")}
+            onPress={() => router.replace("/reminder/new")}
           />
           {/* <MenuOption
             text="🗓️ Add more schedules"
