+++ components/welcome/Notes2.tsx
@@ -1,9 +1,5 @@
 import {
   findNodeHandle,
-  Keyboard,
-  Platform,
-  ScrollView,
-  TouchableWithoutFeedback,
 } from "react-native";
 import { VStack } from "../ui/vstack";
 import { Button, ButtonIcon, ButtonText } from "../ui/button";
