+++ lib/db.ts
@@ -1,5 +1,5 @@
 import * as SQLite from "expo-sqlite";
-import { DB_FILENAME } from "./lib.constants";
+export const DB_FILENAME = "reminders.db";
 
 const db = SQLite.openDatabaseSync(DB_FILENAME);
 export default db;
