--- lib/lib.constants.ts
@@ -1,1 +1,0 @@
-export const DB_FILENAME = "reminders.db";
