+++ lib/notifications/notifications.constants.ts
@@ -4,3 +4,9 @@ export const MAX_ITERATION_LIMIT = 10000;
 export const AMOUNT_TO_SCHEDULE = 64;
 export const UTC_DATE_FORMAT = "YYYY-MM-DD HH:mm:ssZ";
 export const LOCAL_DATE_FORMAT = "YYYY-MM-DD hh:mmZ";
+export const STATUS_COLOR_MAP = {
+  done: "success",
+  no_response: undefined,
+  skip: "info",
+  later: "info",
+} as const;
