+++ lib/reminders/reminders.constants.ts
@@ -1,0 +1,25 @@
+export const NO_REMINDERS_DUE_TEXT = [
+  "Nothing's Due—You're Doing Great! 👍",
+  "No Reminders. Just Vibes. 😎",
+  "You're All Caught Up! 🎉",
+  "You're Good to Go. 👍",
+  "Your Brain Called—You're Free. 🧠",
+  "Stillness is Progress. 🧘",
+  "Rest is Part of the Routine. 🌱",
+  "The Queue is Clear. 💪🏼",
+  "Mission: Completed. 🚀",
+  "Enjoy the downtime. 💤",
+  "You're on break. 🍩",
+  "Inbox: Zero. 📧",
+  "Pause and appreciate. 🙏",
+  "Let yourself recharge. 🔋",
+  "Room to breathe. 💨",
+  "Balance: Restored. 🧘‍♀️",
+  "The day is yours. 🫵",
+  "No loose ends. 👌",
+] as const;
+
+export const REMINDER_LIMIT = {
+  task: 1,
+  recurring: 2,
+} as const;
