+++ lib/revenue-cat/revenue-cat.constants.ts
@@ -1,2 +1,2 @@
+export const UNLIMITED_ENTITLEMENT_NAME = "Unlimited";
 export const UNLIMITED_ENTITLEMENT_ID = "com.offcueapps.offcue.Unlimited";
-export const ALL_DAY_LABEL = "all day";
