+++ lib/revenue-cat/revenue-cat.service.ts
@@ -1,11 +1,12 @@
 import RevenueCatUI, { PAYWALL_RESULT } from "react-native-purchases-ui";
-import { $hasUnlimited } from "../stores/revenueCat";
-import { UNLIMITED_ENTITLEMENT_ID } from "./utils.constants";
+import { UNLIMITED_ENTITLEMENT_ID } from "./revenue-cat.constants";
+import { $hasUnlimited } from "./revenue-cat.store";
 
 export async function presentUnlimitedPaywall() {
-  const paywallResult: PAYWALL_RESULT = await RevenueCatUI.presentPaywallIfNeeded({
-    requiredEntitlementIdentifier: UNLIMITED_ENTITLEMENT_ID,
-  });
+  const paywallResult: PAYWALL_RESULT =
+    await RevenueCatUI.presentPaywallIfNeeded({
+      requiredEntitlementIdentifier: UNLIMITED_ENTITLEMENT_ID,
+    });
 
   switch (paywallResult) {
     case PAYWALL_RESULT.NOT_PRESENTED:
@@ -15,10 +16,10 @@ export async function presentUnlimitedPaywall() {
       return false;
     case PAYWALL_RESULT.PURCHASED:
     case PAYWALL_RESULT.RESTORED:
-        $hasUnlimited.set(true);
-        return true;
+      $hasUnlimited.set(true);
+      return true;
     default:
-        $hasUnlimited.set(false);
-        return false;
+      $hasUnlimited.set(false);
+      return false;
   }
 }
