--- lib/stores/stores.constants.ts
@@ -1,1 +1,0 @@
-export const UNLIMITED_ENTITLEMENT_NAME = "Unlimited";
