+++ lib/utils/format.ts
@@ -1,7 +1,6 @@
 import dayjs from "dayjs";
 import { IntervalType } from "../reminders/reminders.types";
-import { InsertSchedule, Schedule } from "../schedules/schedules.types";
-import { ALL_DAY_LABEL } from "./utils.constants";
+import { InsertSchedule } from "../schedules/schedules.types";
 
 export function chunkIntoPairs<T>(list: T[]) {
   const result = [];
@@ -27,17 +26,9 @@ export function formatFrequencyString(
   }`;
 }
 
-type DayList = (
-  | "monday"
-  | "tuesday"
-  | "wednesday"
-  | "thursday"
-  | "friday"
-  | "saturday"
-  | "sunday"
-)[];
+type DayList = Partial<typeof WEEK_DAYS>;
 
-const weekDays = [
+const WEEK_DAYS = [
   "sunday",
   "monday",
   "tuesday",
@@ -47,7 +38,7 @@ const weekDays = [
   "saturday",
 ] as const;
 
-const DAY_LABEL_MAP: Record<(typeof weekDays)[number], string> = {
+const DAY_LABEL_MAP: Record<(typeof WEEK_DAYS)[number], string> = {
   sunday: "Su",
   monday: "M",
   tuesday: "Tu",
@@ -55,7 +46,9 @@ const DAY_LABEL_MAP: Record<(typeof weekDays)[number], string> = {
   thursday: "Th",
   friday: "F",
   saturday: "Sa",
-};
+} as const;
+
+export const ALL_DAY_LABEL = "all day";
 
 function formatTime(t: string) {
   const formatted = dayjs(`2000-01-01 ${t}`).format("h:mma");
@@ -69,16 +62,16 @@ function formatTime(t: string) {
  * becomes "F-Tu".
  */
 function compressDays(days: DayList): string {
-  const daySet = new Set(days.map((d) => d.toLowerCase()));
+  const daySet = new Set(days.map((d) => d?.toLowerCase()));
   if (daySet.size === 0) return "";
 
   // Find segments in the week array where the day is selected.
   const segments: { start: number; end: number }[] = [];
   let i = 0;
-  while (i < weekDays.length) {
-    if (daySet.has(weekDays[i])) {
+  while (i < WEEK_DAYS.length) {
+    if (daySet.has(WEEK_DAYS[i])) {
       const start = i;
-      while (i < weekDays.length && daySet.has(weekDays[i])) {
+      while (i < WEEK_DAYS.length && daySet.has(WEEK_DAYS[i])) {
         i++;
       }
       segments.push({ start, end: i - 1 });
@@ -92,7 +85,7 @@ function compressDays(days: DayList): string {
   if (segments.length > 1) {
     const firstSeg = segments[0];
     const lastSeg = segments[segments.length - 1];
-    if (firstSeg.start === 0 && lastSeg.end === weekDays.length - 1) {
+    if (firstSeg.start === 0 && lastSeg.end === WEEK_DAYS.length - 1) {
       // Merge the segments: the new range goes from the start of the last segment
       // to the end of the first segment.
       segments[0] = { start: lastSeg.start, end: firstSeg.end };
@@ -103,10 +96,10 @@ function compressDays(days: DayList): string {
   // Format each segment.
   const formattedSegments = segments.map((seg) => {
     if (seg.start === seg.end) {
-      return DAY_LABEL_MAP[weekDays[seg.start]];
+      return DAY_LABEL_MAP[WEEK_DAYS[seg.start]];
     } else {
-      return `${DAY_LABEL_MAP[weekDays[seg.start]]}-${
-        DAY_LABEL_MAP[weekDays[seg.end]]
+      return `${DAY_LABEL_MAP[WEEK_DAYS[seg.start]]}-${
+        DAY_LABEL_MAP[WEEK_DAYS[seg.end]]
       }`;
     }
   });
@@ -123,10 +116,10 @@ export function formatScheduleString(schedule: InsertSchedule) {
     schedule.is_thursday && "thursday",
     schedule.is_friday && "friday",
     schedule.is_saturday && "saturday",
-  ].filter((d) => !!d) as DayList;
+  ].filter((d) => !!d) as unknown as DayList;
 
   // Create a set of all days for checking if every day is included.
-  const inputSet = new Set(days.map((d) => d.toLowerCase()));
+  const inputSet = new Set(days.map((d) => d?.toLowerCase()));
 
   let timeString = `${formatTime(schedule.start_time)} - ${formatTime(
     schedule.end_time
@@ -135,7 +128,7 @@ export function formatScheduleString(schedule: InsertSchedule) {
   if (schedule.start_time === schedule.end_time) timeString = ALL_DAY_LABEL;
 
   // If every day of the week is selected, return only the time string.
-  if (weekDays.every((d) => inputSet.has(d))) {
+  if (WEEK_DAYS.every((d) => inputSet.has(d))) {
     return timeString;
   }
